Fix: bd-key-value 32-bit Android compilation (#372)
* Fix: bd-key-value 32-bit Android compilation

* Update lib.rs

* PR feedback

+++ bd-key-value/src/lib.rs
@@ -168,9 +168,9 @@ impl Store {
               anyhow::bail!("stored string is too short to contain length prefix")
             };
 
-            let len = usize::from_le_bytes(*len_bytes);
+            let len = u64::from_le_bytes(*len_bytes);
 
-            if rest.len() != len {
+            if rest.len() as u64 != len {
               anyhow::bail!("stored string length prefix does not match actual length");
             }
